Revert "Merge conflicts"
This reverts commit e221ce00a7402a35337e76d4d4d283f9f49846ba, reversing
changes made to 35cdf789d87878ad6521a6777bdef6ed5ac27265.
