+++ src/routes/root.tsx
@@ -1,91 +1,625 @@
-import { useMemo, useState } from 'react';
-import { Outlet, useLoaderData } from 'react-router';
-import { Problem, ProblemSessionStats } from '../types';
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { Outlet, useLoaderData, useNavigate, useLocation } from 'react-router';
+import { Link } from 'react-router-dom';
+import { BLANK_CONTRACT, BlogPost, ContractData, ContractProgress, Problem, ProblemSessionStats, Submission } from '../types';
 import { supabase } from '../supabaseClient';
-import { Box, Stack } from '@mui/joy';
-
+import { type Session } from '@supabase/supabase-js';
+import { List as ListIcon } from '@phosphor-icons/react';
+import { Typography, Box, Stack, Drawer, ModalClose, DialogTitle, DialogContent, Button, Option, Select } from '@mui/joy';
+import CategoryList from '../components/CategoryList';
+import whitePaw from '../assets/white_paw.webp';
+import whitePawHover from '../assets/white_paw_hover.webp';
+import logo from '../assets/coding-cat.png';
+import ProblemList from '../components/ProblemList';
+import CustomSearch from '../components/ProblemSearch';
 import getProblemSet from '../utils/getProblemSet';
+import PasswordProtected from './PasswordProtected';
+import { ALL_PFPS } from '../components/profile/UserInfo';
+import ProfileAvatar from '../components/profile/ProfileAvatar';
+import BlogList from '../components/BlogList';
+import getBlogPosts from '../utils/getBlogPosts';
 
-import useAuth from '../hooks/useAuth';
-import useSessionManagement from '../hooks/useSessionManagement';
-import useSearchAndFilter from '../hooks/useSearchAndFilter';
-import useContractData from '../hooks/useContractData';
+interface UserData{
+  name: string,
+  pfp_id: number
+}
 
-import MainLayout from '../components/layout/MainLayout';
-import SidebarDrawer from '../components/layout/SidebarDrawer';
-import UpperNavBar from '../components/layout/UpperNavBar';
-import AppHeader from '../components/layout/AppHeader';
-
+// TODO: This component is huge and should be broken down into smaller components.
+// The main thing that needs to be done is putting the `Drawer` component into its own separate file.
 export default function App() {
+  const navigate = useNavigate();
+  const location = useLocation();
+  const [session, setSession] = useState<Session | null>(null);
+  const [userData, setUserData] = useState<UserData | null>(null);
+  const [activeSession, setActiveSession] = useState(false);
+  const [sessionId, setSessionId] = useState<string | null>(null);
+  const [sessionDuration, setSessionDuration] = useState<number>(0);
+  const [sessionStartTime, setSessionStartTime] = useState<Date | null>(null);
+  const [sessionRemainingSeconds, setSessionRemainingSeconds] = useState<number>(0);
+  const [plannedExerciseCount, setPlannedExerciseCount] = useState<number>(0);
+  const sessionTimerRef = useRef<NodeJS.Timeout | null>(null);
+  const [isRecoverySession, setIsRecoverySession] = useState(false);
+  const [isAdmin, setIsAdmin] = useState(false);
+  const [drawerOpen, setDrawerOpen] = useState(false);
+  const [openCategory, setOpenCategory] = useState(false);
+  const [activeProblem, setActiveProblem] = useState<null | string>(null);
   const problems = useLoaderData() as Problem[];
+  const allCategories = useMemo(() => {
+    const categories = problems
+      .map((c) => c.meta.category)
+      .filter((c, index, array) => array.indexOf(c) === index)
+      .sort((a, b) => a.localeCompare(b));
 
-  const { session, userData, isAdmin, isRecoverySession, fetchProfile } = useAuth();
-  const { progress, contractProgress, fetchProgress } = useContractData(session);
-  const {
-    activeSession, sessionId, sessionDuration,sessionRemainingSeconds,plannedExerciseCount,
-    sessionTimerRunning, endSession,formatTime, } = useSessionManagement(session);
-
+    const specialCategories: string[] = [];
+    if(problems.some(p => p.meta.question_type[0] === "haystack")) specialCategories.push("haystack");
+    if(problems.some(p => p.meta.question_type[0] === "mutation")) specialCategories.push("mutation");
+  
+    return [...categories, ...specialCategories];
+  }, [problems]);
+  const [activeCategory, setActiveCategory] = useState<string | null>(() => {return 'Fundamentals';});
+  const [query, setQuery] = useState("");
+  const [difficulty, setDifficulty] = useState("");
+  const [searchedProblems, setSearchedProblems] = useState<Problem[]>([]);
+  const [searchedBlogs, setSearchedBlogs] = useState<BlogPost[]>([]);
+  const [selectedTab, setSelectedTab] = useState("");
+  const [contract, setContract] = useState<ContractData>(BLANK_CONTRACT);
+  const [progress, setProgress] = useState<Submission[]>([]);
   const [problemSessionStats, setProblemSessionStats] = useState<Record<string, ProblemSessionStats>>({});
+  const [sessionTimerRunning, setSessionTimerRunning] = useState(false);
 
-  const search = useSearchAndFilter(problems);
+  const [keyboardSelectedProblem, setKeyboardSelectedProblem] = useState(activeProblem);
+  const [keyboardSelectedCategory, setKeyboardSelectedCategory] = useState(activeCategory);
 
-  const {
-    query, setQuery, difficulty, setDifficulty, activeCategory, activeProblem,
-    setActiveProblem, drawerOpen, setDrawerOpen, openCategory, setOpenCategory, selectedTab,
-    setSelectedTab, searchedProblems, searchedBlogs, handleSelectedCategory,
-    handleSelectedProblem, keyboardSelectedProblem, keyboardSelectedCategory
-  } = search;
+  const contractProgress: ContractProgress = contract.Coding.problemsToSolveByCategory;
+  contractProgress["mutation"] = contract.Mutation.problemsToSolve;
+  contractProgress["haystack"] = contract.Haystack.problemsToSolve;
+
+  let newDifficulty = difficulty;
+  if (newDifficulty === "all") newDifficulty = "";
+
+  const filteredProblems = useMemo(() => {
+    return problems.filter(problem => {
+      return problem.meta.title.toLowerCase().includes(query.toLowerCase().trim()) &&
+        problem.meta.difficulty.includes(newDifficulty);
+    });
+  }, [problems, query, newDifficulty])
+
+  useEffect(() => {
+    setSearchedProblems(filteredProblems);
+  }, [filteredProblems])
+
+  
+  useEffect(() => {
+    (async () => {
+      const posts = await getBlogPosts();
+      setSearchedBlogs(posts);
+    })();
+  })
+
+  function handleSelectedCategory(category: string){
+    setActiveCategory(category)
+    setActiveProblem(null)
+    setOpenCategory(false)
+    if(category === "coding") setSelectedTab("") 
+    else setSelectedTab("List")
+  }
+
+  function handleSelectedProblem(name: string){
+    setActiveProblem(name)
+    setDrawerOpen(false)
+  }
+
+  useEffect(() => {
+    supabase.auth.onAuthStateChange((_event, session) => {
+      if(_event === 'PASSWORD_RECOVERY'){
+        setIsRecoverySession(true);
+        return;
+      }
+      setIsRecoverySession(false);
+      setSession(session);
+      if (session?.user) {
+        supabase
+          .from('profiles')
+          .select('is_admin')
+          .eq('profile_id', session.user.id)
+          .single()
+          .then(({ data, error }) => {
+            if (!error && data) {
+              setIsAdmin(data.is_admin);
+            }
+          });
+      } else {
+        setIsAdmin(false);
+      }
+    });
+    if(!isRecoverySession){
+      supabase.auth.getSession().then(({ data: { session } }) => {
+        setSession(session);
+        if (session?.user){
+          supabase
+            .from('profiles')
+            .select('is_admin')
+            .eq('profile_id', session.user.id)
+            .single()
+            .then(({ data, error }) => {
+              if (!error && data) {
+                setIsAdmin(data.is_admin);
+              }
+            });
+        }
+      });
+    }
+  }, [isRecoverySession]);
+
+  useEffect(() => {
+    (async () => {
+      if (!session) return;
+
+      const { data } = await supabase
+        .from("contracts")
+        .select("data")
+        .eq("profile_id", session?.user.id)
+        .order("updated_at", { ascending: false })
+        .limit(1);
+
+      if (data?.[0]?.data) setContract(data[0].data);
+      else setContract(BLANK_CONTRACT)
+    })();
+  }, [session]);
+
+  const fetchProfile = useCallback(async () => {
+    if (!session) return;
+    const { user } = session;
+
+    const { data } = await supabase
+      .from('profiles')
+      .select('username, pfp_id')
+      .eq('profile_id', user.id)
+      .single();
+
+    if (data) {
+      setUserData({
+        name: data.username,
+        pfp_id: data.pfp_id
+      });
+    }
+  }, [session]);
+
+  useEffect(() => {
+    fetchProfile();
+  }, [fetchProfile]);
+
+  const fetchProgress = useCallback(async () => {
+    if(!session) return;
+    const { data: submissions, error } = await supabase
+      .from('submissions')
+      .select('problem_title, passed_tests, total_tests, question_type')
+      .eq('profile_id', session.user.id);
+    if(!error) setProgress(submissions || []);
+  }, [session]);
+
+  useEffect(() => {
+    fetchProgress();
+  }, [fetchProgress]);
 
   const problemListProps = {
-    selectedTab, setSelectedTab, searchedProblems, selectedCategory: activeCategory,
-    activeProblem, onSelectProblem: handleSelectedProblem, closeDrawer: () => setDrawerOpen(false),
-    session, contractProgress, progress
+    selectedTab,
+    setSelectedTab,
+    searchedProblems,
+    selectedCategory: activeCategory,
+    activeProblem,
+    onSelectProblem: handleSelectedProblem,
+    closeDrawer: () => setDrawerOpen(false),
+    session,
+    contractProgress,
+    progress,
+    keyboardSelected: keyboardSelectedProblem,
   };
-
+  
   const blogListProps = {
-    searchedBlogs: searchedBlogs, selectedTab, setSelectedTab, selectedCategory: activeCategory,
-    activeBlog: activeProblem, closeDrawer: () => setDrawerOpen(false),
+    searchedBlogs: searchedBlogs,
+    selectedTab,
+    setSelectedTab,
+    selectedCategory: activeCategory,
+    activeBlog: activeProblem,
+    closeDrawer: () => setDrawerOpen(false)
   };
 
-  const signOut = async () => {
-    await supabase.auth.signOut();
+  const startSession = (sessionIdFromState: string, durationMinutes: number, exerciseCount: number) => {
+    setSessionId(sessionIdFromState);
+    setSessionDuration(durationMinutes);
+    setPlannedExerciseCount(exerciseCount);
+    setSessionStartTime(new Date());
+    setSessionRemainingSeconds(durationMinutes * 60);
+    setActiveSession(true);
+    setSessionTimerRunning(true);
   };
+
+  const endSession = () => {
+    if (sessionTimerRef.current) clearInterval(sessionTimerRef.current);
+    setActiveSession(false);
+    setSessionId(null);
+    setSessionStartTime(null);
+    setSessionDuration(0);
+    setSessionRemainingSeconds(0);
+    setProblemSessionStats({});
+    setSessionTimerRunning(false);
+  };
+
+  const formatTime = (seconds: number): string => {
+    const minutes = Math.floor(seconds / 60);
+    const remainingSeconds = seconds % 60;
+    return `${minutes.toString().padStart(2, '0')}:${remainingSeconds.toString().padStart(2, '0')}`;
+  };
+
+  // Handle session start when coming back from PreSessionForm
+  useEffect(() => {
+    const locationState = location.state as any;
+    const sessionIdFromState = locationState?.sessionId;
+    const fromPreSession = locationState?.fromPreSession;
+
+    if (location.pathname !== '/' || !fromPreSession || !session?.user) return;
+
+    const fetchSessionData = async () => {
+      const { data, error } = await supabase
+        .from('sessions')
+        .select('planned_duration_minutes, exercise_goals')
+        .eq('id', sessionIdFromState)
+        .eq('profile_id', session.user.id)
+        .single();
+
+      if (!error && data) {
+        startSession(sessionIdFromState, data.planned_duration_minutes, data.exercise_goals);
+      }
+    };
+    fetchSessionData();
+  }, [location.pathname, location.state, session?.user, navigate]);
+
+  // Handle session reset when coming back from PostSessionForm
+  useEffect(() => {
+    if (location.pathname === '/' && activeSession && (location.state as any)?.fromPostSession) {
+      endSession();
+    }
+  }, [location.pathname, location.state, activeSession]);
+
+  // Session countdown timer
+  useEffect(() => {
+    if (!activeSession || !sessionStartTime) return;
+
+    sessionTimerRef.current = setInterval(() => {
+      const elapsed = Math.floor((new Date().getTime() - sessionStartTime.getTime()) / 1000);
+      const remaining = Math.max(0, sessionDuration * 60 - elapsed);
+      setSessionRemainingSeconds(remaining);
+
+      if (remaining <= 0) {
+        clearInterval(sessionTimerRef.current!);
+        setActiveSession(false);
+        setSessionStartTime(null);
+        setSessionRemainingSeconds(0);
+        setSessionTimerRunning(false);
+        navigate('/post-session', {
+          state: {
+            sessionId,
+            timerExpired: true
+          }
+        });
+      }
+    }, 1000);
+
+    return () => {
+      if (sessionTimerRef.current) clearInterval(sessionTimerRef.current);
+    };
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [activeSession, sessionStartTime, sessionDuration]);
+
+  // keybinds navigation
+  const handleKeyPress = useCallback((event: KeyboardEvent) => {
+    // Ctrl + D opens Drawer
+    if(event.ctrlKey && event.key === "d"){
+      event.preventDefault();
+      setDrawerOpen(o => !o);
+    }
+
+    if(drawerOpen){
+      if(openCategory){
+        // up / down selects category
+        if(event.key === "ArrowUp"){
+          event.preventDefault();
+          const currentIndex = allCategories.indexOf(keyboardSelectedCategory ?? "");
+          const prevIndex = currentIndex <= 0 
+            ? allCategories.length - 1
+            : currentIndex - 1;
+          setKeyboardSelectedCategory(allCategories[prevIndex]);
+        }
+        if(event.key === "ArrowDown"){
+          event.preventDefault();
+          const currentIndex = allCategories.indexOf(keyboardSelectedCategory ?? "");
+          const nextIndex = currentIndex >= allCategories.length - 1
+            ? 0
+            : currentIndex + 1;
+          setKeyboardSelectedCategory(allCategories[nextIndex]);
+        }
+
+        // select category
+        if (event.key === "Enter") {
+          event.preventDefault();
+          if (keyboardSelectedCategory) handleSelectedCategory(keyboardSelectedCategory);
+        }
+      }
+      else{
+        // filters problems in active category
+        const categoryProblems = searchedProblems
+          .filter(p => {
+            const questionType = p.meta.question_type[0];
+            const cat = questionType === "coding"
+              ? p.meta.category
+              : questionType;
+            return cat === activeCategory;
+          })
+          .map(p => p.meta.name);
+        
+        // up / down selects problem
+        if(event.key === "ArrowUp"){
+          event.preventDefault();
+          const currentIndex = categoryProblems.indexOf(keyboardSelectedProblem ?? "");
+          const prevIndex = currentIndex <= 0 
+            ? categoryProblems.length - 1 
+            : currentIndex - 1;
+          setKeyboardSelectedProblem(categoryProblems[prevIndex]);
+        }
+        if(event.key === "ArrowDown"){
+          event.preventDefault();
+          const currentIndex = categoryProblems.indexOf(keyboardSelectedProblem ?? "");
+          const nextIndex = currentIndex >= categoryProblems.length - 1 
+            ? 0
+            : currentIndex + 1;
+          setKeyboardSelectedProblem(categoryProblems[nextIndex]);
+        }
+
+        // select new problem
+        if(event.key === "Enter"){
+          event.preventDefault();
+          if(keyboardSelectedProblem){
+            handleSelectedProblem(keyboardSelectedProblem);
+            navigate(`/problems/${keyboardSelectedProblem}`);
+          }
+        }
+      }
+
+      // left / right opens category list
+      if(event.key === "ArrowLeft"){
+        setKeyboardSelectedCategory(activeCategory);
+        setOpenCategory(true);
+      }
+      if(event.key === "ArrowRight"){
+        setKeyboardSelectedCategory(activeCategory);
+        setOpenCategory(false);
+      }
+    }
+  }, [navigate, keyboardSelectedProblem, drawerOpen, openCategory, allCategories, searchedProblems, activeCategory, keyboardSelectedCategory]);
+
+  // on new category selected, set selectedProblem to first problem
+  useEffect(() => {
+    const first = searchedProblems
+      .filter(p => p.meta.category === activeCategory)
+      .map(p => p.meta.name)[0] ?? null;
+    setKeyboardSelectedProblem(activeProblem ?? first);
+  }, [activeCategory, searchedProblems]);
+
+  useEffect(() => {
+    document.addEventListener('keydown', handleKeyPress);
+    return () => {
+      document.removeEventListener('keydown', handleKeyPress);
+    };
+  }, [handleKeyPress]);
 
   return (
-    <MainLayout openDrawer={() => setDrawerOpen(true)}>
-      <SidebarDrawer
-        drawerOpen={drawerOpen} onDrawerClose={() => setDrawerOpen(false)} openCategory={openCategory}
-        setOpenCategory={setOpenCategory} difficulty={difficulty} setDifficulty={setDifficulty}
-        query={query} setQuery={setQuery} searchedProblems={searchedProblems} searchedBlogs={searchedBlogs}
-        activeCategory={activeCategory} handleSelectedCategory={handleSelectedCategory}
-        problemListProps={problemListProps} blogListProps={blogListProps} selectedTab={selectedTab}
-        setSelectedTab={setSelectedTab} keyboardSelected={keyboardSelectedCategory}
+    <Box sx={{ display:'flex', height: "100%", flex: 1}}>
+      <Stack
+       sx={{
+          width: '6em',
+          cursor: 'pointer',
+          alignItems: 'center',
+          position: 'relative',
+          '&::after': {
+            content: '""',
+            position: 'absolute',
+            top: '50%',
+            width: '70px',
+            height: '70px',
+            backgroundImage: `url(${whitePaw})`,
+            backgroundSize: 'contain',
+            backgroundRepeat: 'no-repeat',
+            backgroundPosition: 'center',
+            transform: 'rotate(-90deg)'
+          },
+          '&:hover::after': {
+            backgroundImage: `url(${whitePawHover})`
+          }
+        }}
+        className="desktop-bar"
+        onClick={() => setDrawerOpen(true)}
       />
-
-      <Stack sx={{ width: '100%' }}>
-        <UpperNavBar
-          openDrawer={() => setDrawerOpen(true)} session={session} isRecoverySession={isRecoverySession}
-          activeSession={activeSession} sessionTimerRunning={sessionTimerRunning} formatTime={formatTime}
-          sessionRemainingSeconds={sessionRemainingSeconds} endSession={endSession} sessionId={sessionId}
-          userData={userData} isAdmin={isAdmin} signOut={signOut}
-        />
-
-        <AppHeader />
-
+      <Drawer
+        open={drawerOpen}
+        onClose={() => setDrawerOpen(false)}
+        size="lg"
+        // Temporary fix for: https://github.com/coding-cat-official/coding-cat/pull/56
+        sx={{
+          "--ModalClose-inset": "1rem",
+          "--Drawer-verticalSize": "clamp(500px, 60%, 100%)",
+          "--Drawer-horizontalSize": "100vw",
+          "--Drawer-titleMargin": "1rem 1rem calc(1rem / 2)",
+        }}
+      >
+        <ModalClose />
+          <Stack width="100%" direction="row" justifyContent="space-between" padding={'10px'} className="big-navbar" sx={{alignItems: "center"}}>
+            <DialogTitle level='h1'  sx={{ fontFamily: '"Silkscreen", monospace', padding: "5px", fontSize: "30pt"}}>
+              Coding Cat
+            </DialogTitle>
+            <Stack marginRight="5em" direction="row" gap={3} className="problemList-search-filter">
+              <Select sx={{ width: "150px", fontWeight: "normal", fontFamily: "Silkscreen" }} placeholder="Difficulty" value={difficulty} onChange={(e, newValue) => setDifficulty(newValue || "")}>
+                <Option sx={{fontFamily: "Silkscreen"}} value="all">All</Option>
+                <Option sx={{fontFamily: "Silkscreen"}} value="easy">Easy</Option>
+                <Option sx={{fontFamily: "Silkscreen"}} value="medium">Medium</Option>
+                <Option sx={{fontFamily: "Silkscreen"}} value="hard">Hard</Option>
+              </Select>
+              <CustomSearch query={query} setQuery={setQuery} placeholder="Search for exercises..." />
+            </Stack>
+          </Stack>
+        <DialogContent>
+          <Box sx={{ display: 'flex', overflow: 'hidden', gap: "16px" }}>
+              <Button className="mobile-categoryList" onClick={() => setOpenCategory(true)}>&gt;</Button>
+              <Drawer open={openCategory} onClose={() => setOpenCategory(false)} sx={{ flex: 1, width: 300, overflowY: 'auto',}} className="mobile-categoryList">
+                <CategoryList
+                  searchedProblems={searchedProblems}
+                  activeCategory={activeCategory}
+                  onSelectCategory={handleSelectedCategory}
+                  session={session}
+                  contractProgress={contractProgress}
+                  keyboardSelected={keyboardSelectedCategory}
+                />
+              </Drawer>
+              <Box sx={{ flex: 1, width: 300, overflowY: 'auto',}} className="categoryList">
+                <CategoryList
+                  searchedProblems={searchedProblems}
+                  activeCategory={activeCategory}
+                  onSelectCategory={handleSelectedCategory}
+                  session={session}
+                  contractProgress={contractProgress}
+                  keyboardSelected={keyboardSelectedCategory}
+                />
+              </Box>
+              <Box sx={{ flex: 3}} className="parent-problemList">
+                {activeCategory === 'test-questions' ? (
+                  <PasswordProtected {...problemListProps}/>
+                ) : activeCategory === 'blogs' ? (
+                  <BlogList {...blogListProps}/>
+                ): (
+                  <ProblemList {...problemListProps} />
+                )}
+              </Box>
+            </Box>
+        </DialogContent>
+      </Drawer>
+      <Stack
+        className= 'main'
+        direction="column"
+        sx={{
+          width: '100%',
+          minHeight: "100%",
+          height: "100%", 
+          justifyContent: "start",
+          alignItems: "center",
+          overflowY: "scroll", 
+          position:'relative' 
+        }} >
+          <Stack sx={{ width: '100%', display: 'flex', flexDirection: 'row'}} className="upper-nav">
+            <Button sx={{ margin: '10px 10px 0 10px', cursor: 'pointer'}} onClick={() => setDrawerOpen(true)} className="mobile-bar">
+              { /* TODO: get non-deprecated icon */ }
+              <ListIcon size={20} />
+            </Button>
+            <Box sx={{ margin: '10px 10px 0 10px', display: 'flex', gap: 1 }} className="account-btns">
+              {session && !isRecoverySession ? (
+                <>
+                  <Button
+                    onClick={() => {
+                      if (activeSession && sessionTimerRunning) {
+                        endSession();
+                        navigate('/post-session', { state: { sessionId } });
+                      } else if (activeSession && !sessionTimerRunning) {
+                        navigate('/post-session', { state: { sessionId } });
+                      }else {
+                        navigate('/session');
+                      }
+                    }}
+                    sx={{
+                      backgroundColor: activeSession ? '#d4ff99' : '#d4ff99',
+                      color: '#1a3e00',
+                      borderRadius: '999px',
+                      px: 2,
+                      py: 1,
+                      minWidth: '240px',
+                      boxShadow: '0 4px 10px rgba(0,0,0,0.08)',
+                      '&:hover': {
+                        backgroundColor: '#c7f68e',
+                      }
+                    }}
+                    title={activeSession ? "Click to end session" : "Start a new session"}
+                  >
+                  {sessionTimerRunning
+                    ? `Ongoing session — ${formatTime(sessionRemainingSeconds)} left`
+                    : activeSession
+                      ? 'Complete Session Reflection'
+                      : 'Start Session'}
+                  </Button>
+                  <Link to="/profile">
+                    <Button>
+                      <Stack flexDirection="row" alignItems="center" gap={1}>
+                        {
+                          userData?.pfp_id != null ?
+                            <ProfileAvatar 
+                              fileName={ALL_PFPS[userData.pfp_id]}
+                              height={25}
+                              width={25}
+                            />
+                          : <></>
+                        }
+                        { userData?.name ?? "Profile" }
+                      </Stack>
+                    </Button>
+                  </Link>
+                  <Button onClick={() => supabase.auth.signOut()}>Sign Out</Button>
+                  {isAdmin && (
+                    <Link to="/admin">
+                      <Button color="warning">Admin</Button>
+                    </Link>
+                  )}
+                </>
+              ) : (
+                <>
+                  <Link to="/signin">
+                    <Button>Login</Button>
+                  </Link>
+                  <Link to="/register">
+                    <Button>Register</Button>
+                  </Link>
+                </>
+              )}
+            </Box>
+          </Stack>
+        
+        <Stack sx={{ width: '100%' }} direction="row" alignItems="center" justifyContent="center"  className="logo">
+          <Link to="/">
+            <Box component="img" src={logo} sx={{ maxHeight: "80px", marginTop: "5px", marginRight:"15px" }}/>
+          </Link>
+          <Typography  sx={{ fontFamily: '"Silkscreen", monospace', fontSize: "35pt"}} level="h1">
+            Coding Cat!
+          </Typography>
+        </Stack>
+        
         <Box width="100%" height="100%">
-          <Outlet
-            context={{
-              setActiveProblem, session, isAdmin, refetchProgress: fetchProgress,
-              refetchProfile: fetchProfile, activeSession, sessionId, sessionRemainingSeconds,
-              sessionDuration, plannedExerciseCount, problemSessionStats, setProblemSessionStats,
-              progress, sessionTimerRunning,
-            }}
-          />
+          <Outlet context={{ 
+            setActiveProblem, 
+            session, 
+            isAdmin, 
+            refetchProgress: fetchProgress,
+            refetchProfile: fetchProfile,
+            activeSession,
+            sessionId,
+            sessionRemainingSeconds,
+            sessionDuration,
+            plannedExerciseCount,
+            problemSessionStats,
+            setProblemSessionStats,
+            progress,
+            sessionTimerRunning
+          }} />
         </Box>
+        
       </Stack>
-    </MainLayout>
-  );
-}
+    </Box>
+)};
 
 export async function problemListLoader(): Promise<Problem[]> {
   return await getProblemSet() as Problem[];
